Consolidate unit tests and simplify Makefile test targets
- Consolidated 10 test files down to 6 files (40% reduction)
- Deleted redundant test files and merged into core test modules:
  - test_exception_handling.py → test_ray_manager.py
  - test_streaming_logs.py → test_ray_manager.py
  - test_state_management.py → test_ray_manager.py
  - test_utils.py → conftest.py (utilities moved)
- Enhanced test_ray_manager.py as main consolidation target
- Streamlined test_integration.py and test_multi_node_cluster.py
- Consolidated worker tests into test_worker_manager.py
- Improved conftest.py with shared fixtures and utilities

Makefile simplification:
- Reduced test targets from 8 to 3 simple targets:
  - make test (default, complete test suite)
  - make test-fast (excludes e2e tests)
  - make test-e2e (e2e tests with cleanup)
- Removed: test-full, test-tool-functions, test-e2e-verbose, etc.
- Updated help documentation

Testing verified:
- All 52 tests still pass
- Coverage maintained at 56%
- No functionality lost in consolidation
- Faster test execution due to reduced overhead
