+++ tests/test_integration.py
@@ -14,28 +14,9 @@ from ray_mcp.ray_manager import RayManager
 from ray_mcp.tool_registry import ToolRegistry
 
 
-def get_text_content(result) -> str:
-    """Extract text content from MCP response."""
-    if isinstance(result, list) and len(result) > 0:
-        if hasattr(result[0], "text"):
-            return result[0].text
-        elif isinstance(result[0], dict) and "text" in result[0]:
-            return result[0]["text"]
-    return str(result)
-
-
-async def call_tool(
-    tool_name: str, arguments: Optional[Dict[str, Any]] = None
-) -> List[TextContent]:
-    """Helper function to call a tool through the registry."""
-    registry = ToolRegistry(RayManager())
-    result = await registry.execute_tool(tool_name, arguments or {})
-    return [TextContent(type="text", text=json.dumps(result, indent=2))]
-
-
 @pytest.mark.integration
 class TestMCPIntegration:
-    """Integration tests for MCP server functionality."""
+    """Essential integration tests for MCP server functionality."""
 
     @pytest.fixture(autouse=True)
     def patch_ray_manager(self):
@@ -44,8 +25,89 @@ class TestMCPIntegration:
             yield
 
     @pytest.mark.asyncio
-    async def test_list_tools_returns_all_tools(self):
-        """Test that list_tools returns all available tools."""
+    async def test_complete_workflow_cluster_and_job_lifecycle(self):
+        """Test complete workflow: cluster init -> job submit -> job inspect -> job cancel -> cluster stop."""
+        registry = ToolRegistry(RayManager())
+
+        # Mock RayManager methods for complete workflow
+        with patch.object(registry.ray_manager, "init_cluster") as mock_init:
+            with patch.object(registry.ray_manager, "submit_job") as mock_submit:
+                with patch.object(registry.ray_manager, "inspect_job") as mock_inspect:
+                    with patch.object(registry.ray_manager, "cancel_job") as mock_cancel:
+                        with patch.object(registry.ray_manager, "stop_cluster") as mock_stop:
+                            # Setup mock responses
+                            mock_init.return_value = {
+                                "status": "connected",
+                                "message": "Successfully connected to Ray cluster at 127.0.0.1:10001",
+                                "cluster_address": "127.0.0.1:10001",
+                                "dashboard_url": "http://127.0.0.1:8265",
+                                "node_id": "node_123",
+                                "job_client_status": "ready",
+                            }
+                            mock_submit.return_value = {
+                                "status": "success",
+                                "result_type": "submitted",
+                                "job_id": "job_123",
+                            }
+                            mock_inspect.return_value = {
+                                "status": "success",
+                                "job_id": "job_123",
+                                "job_status": "RUNNING",
+                            }
+                            mock_cancel.return_value = {
+                                "status": "success",
+                                "result_type": "cancelled",
+                                "job_id": "job_123",
+                            }
+                            mock_stop.return_value = {
+                                "status": "stopped",
+                                "message": "Ray cluster stopped successfully",
+                            }
+
+                            # Mock Ray availability and initialization for job operations
+                            with patch("ray_mcp.ray_manager.RAY_AVAILABLE", True):
+                                with patch("ray_mcp.ray_manager.ray") as mock_ray:
+                                    mock_ray.is_initialized.return_value = True
+                                    registry.ray_manager._update_state(initialized=True)
+
+                                    # Test complete workflow
+                                    # 1. Initialize cluster
+                                    result = await registry.execute_tool("init_ray", {"num_cpus": 4})
+                                    assert result["status"] == "connected"
+                                    mock_init.assert_called_once()
+
+                                    # 2. Submit job
+                                    result = await registry.execute_tool(
+                                        "submit_job", {"entrypoint": "python test.py"}
+                                    )
+                                    assert result["status"] == "success"
+                                    assert result["result_type"] == "submitted"
+                                    assert "job_id" in result
+                                    mock_submit.assert_called_once()
+
+                                    # 3. Inspect job
+                                    result = await registry.execute_tool(
+                                        "inspect_job", {"job_id": "job_123"}
+                                    )
+                                    assert result["status"] == "success"
+                                    mock_inspect.assert_called_once()
+
+                                    # 4. Cancel job
+                                    result = await registry.execute_tool(
+                                        "cancel_job", {"job_id": "job_123"}
+                                    )
+                                    assert result["status"] == "success"
+                                    assert result.get("result_type") == "cancelled"
+                                    mock_cancel.assert_called_once()
+
+                                    # 5. Stop cluster
+                                    result = await registry.execute_tool("stop_ray", {})
+                                    assert result["status"] == "stopped"
+                                    mock_stop.assert_called_once()
+
+    @pytest.mark.asyncio
+    async def test_tool_list_and_schema_validation(self):
+        """Test that all tools are available with valid schemas."""
         registry = ToolRegistry(RayManager())
         tools = registry.get_tool_list()
 
@@ -65,116 +127,25 @@ class TestMCPIntegration:
         for expected_tool in expected_tools:
             assert expected_tool in tool_names, f"Missing tool: {expected_tool}"
 
-        # Check that all tools have proper schemas
+        # Check that all tools have valid schemas
         for tool in tools:
             assert hasattr(tool, "inputSchema")
             assert isinstance(tool.inputSchema, dict)
+            assert "type" in tool.inputSchema
+            assert tool.inputSchema["type"] == "object"
 
     @pytest.mark.asyncio
-    async def test_tool_schemas_are_valid(self):
-        """Test that tool schemas are valid JSON schemas."""
-        registry = ToolRegistry(RayManager())
-        tools = registry.get_tool_list()
-
-        for tool in tools:
-            schema = tool.inputSchema
-            assert "type" in schema
-            assert schema["type"] == "object"
-            if "properties" in schema:
-                assert isinstance(schema["properties"], dict)
-
-    @pytest.mark.asyncio
-    async def test_complete_workflow_cluster_management(self):
-        """Test complete workflow for cluster management."""
+    async def test_error_propagation_and_ray_unavailable(self):
+        """Test error handling when Ray is unavailable or tools fail."""
         registry = ToolRegistry(RayManager())
 
-        # Mock RayManager methods
-        with patch.object(registry.ray_manager, "init_cluster") as mock_init:
-            with patch.object(registry.ray_manager, "stop_cluster") as mock_stop:
-                mock_init.return_value = {
-                    "status": "connected",
-                    "message": "Successfully connected to Ray cluster at 127.0.0.1:10001",
-                    "cluster_address": "127.0.0.1:10001",
-                    "dashboard_url": "http://127.0.0.1:8265",
-                    "node_id": "node_123",
-                    "job_client_status": "ready",
-                }
-                mock_stop.return_value = {
-                    "status": "stopped",
-                    "message": "Ray cluster stopped successfully",
-                }
+        # Test Ray unavailable
+        with patch("ray_mcp.ray_manager.RAY_AVAILABLE", False):
+            result = await registry.execute_tool("init_ray", {"num_cpus": 4})
+            assert result["status"] == "error"
+            assert "Ray is not available" in result["message"]
 
-                # Test cluster initialization
-                result = await registry.execute_tool("init_ray", {"num_cpus": 4})
-                assert result["status"] == "connected"
-                mock_init.assert_called_once()
-
-                # Test cluster stop
-                result = await registry.execute_tool("stop_ray", {})
-                assert result["status"] == "stopped"
-                mock_stop.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_complete_workflow_job_lifecycle(self):
-        """Test complete workflow for job lifecycle."""
-        registry = ToolRegistry(RayManager())
-
-        # Mock RayManager methods
-        with patch.object(registry.ray_manager, "submit_job") as mock_submit:
-            with patch.object(registry.ray_manager, "inspect_job") as mock_inspect:
-                with patch.object(registry.ray_manager, "cancel_job") as mock_cancel:
-                    mock_submit.return_value = {
-                        "status": "success",
-                        "result_type": "submitted",
-                        "job_id": "job_123",
-                    }
-                    mock_inspect.return_value = {
-                        "status": "success",
-                        "job_id": "job_123",
-                        "job_status": "RUNNING",
-                    }
-                    mock_cancel.return_value = {
-                        "status": "success",
-                        "result_type": "cancelled",
-                        "job_id": "job_123",
-                    }
-
-                    # Mock Ray availability and initialization
-                    with patch("ray_mcp.ray_manager.RAY_AVAILABLE", True):
-                        with patch("ray_mcp.ray_manager.ray") as mock_ray:
-                            mock_ray.is_initialized.return_value = True
-                            registry.ray_manager._update_state(initialized=True)
-
-                            # Test job submission
-                            result = await registry.execute_tool(
-                                "submit_job",
-                                {"entrypoint": "python test.py"},
-                            )
-                            assert result["status"] == "success"
-                            assert result["result_type"] == "submitted"
-                            assert "job_id" in result
-                            mock_submit.assert_called_once()
-
-                            # Test job inspection
-                            result = await registry.execute_tool(
-                                "inspect_job", {"job_id": "job_123"}
-                            )
-                            assert result["status"] == "success"
-                            mock_inspect.assert_called_once()
-
-                            # Test job cancellation
-                            result = await registry.execute_tool(
-                                "cancel_job", {"job_id": "job_123"}
-                            )
-                            assert result["status"] == "success"
-                            assert result.get("result_type") == "cancelled"
-                            mock_cancel.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_error_propagation(self):
-        """Test that errors are properly propagated from RayManager to tool registry."""
-        registry = ToolRegistry(RayManager())
-
+        # Test exception propagation
         with patch.object(registry.ray_manager, "init_cluster") as mock_init:
             mock_init.side_effect = Exception("Ray initialization failed")
 
@@ -183,57 +154,34 @@ class TestMCPIntegration:
             assert "Ray initialization failed" in result["message"]
 
     @pytest.mark.asyncio
-    async def test_ray_unavailable_error(self):
-        """Test error handling when Ray is not available."""
+    async def test_parameter_validation_and_complex_params(self):
+        """Test parameter validation and complex parameter structures."""
         registry = ToolRegistry(RayManager())
 
-        with patch("ray_mcp.ray_manager.RAY_AVAILABLE", False):
-            result = await registry.execute_tool("init_ray", {"num_cpus": 4})
-            assert result["status"] == "error"
-            assert "Ray is not available" in result["message"]
-
-    @pytest.mark.asyncio
-    async def test_parameter_validation_integration(self):
-        """Test parameter validation at the integration level."""
-        registry = ToolRegistry(RayManager())
-
-        # Test with invalid parameters
         with patch.object(registry.ray_manager, "init_cluster") as mock_init:
             mock_init.return_value = {"status": "connected"}
 
-            # Test with valid parameters
-            result = await registry.execute_tool(
-                "init_ray", {"num_cpus": 4, "num_gpus": 1}
-            )
+            # Test complex parameters
+            complex_params = {
+                "num_cpus": 4,
+                "num_gpus": 1,
+                "object_store_memory": 1000000000,
+                "worker_nodes": [
+                    {"num_cpus": 2, "num_gpus": 0, "node_name": "worker-1"},
+                    {"num_cpus": 2, "num_gpus": 1, "node_name": "worker-2"},
+                ],
+                "dashboard_port": 8265,
+                "head_node_host": "127.0.0.1",
+            }
+
+            result = await registry.execute_tool("init_ray", complex_params)
             assert result["status"] == "connected"
 
-            # Verify that parameters were passed correctly
+            # Verify complex parameters were passed correctly
             mock_init.assert_called_once()
             call_args = mock_init.call_args[1]
             assert call_args["num_cpus"] == 4
-            assert call_args["num_gpus"] == 1
-
-    @pytest.mark.asyncio
-    async def test_response_format_consistency(self):
-        """Test that all tool responses have consistent format."""
-        registry = ToolRegistry(RayManager())
-
-        with patch.object(registry.ray_manager, "init_cluster") as mock_init:
-            mock_init.return_value = {
-                "status": "connected",
-                "message": "Successfully connected to Ray cluster at 127.0.0.1:10001",
-                "cluster_address": "127.0.0.1:10001",
-                "dashboard_url": "http://127.0.0.1:8265",
-                "node_id": "node_123",
-                "job_client_status": "ready",
-            }
-
-            result = await registry.execute_tool("init_ray", {"num_cpus": 4})
-
-            # Check response format
-            assert isinstance(result, dict)
-            assert "status" in result
-            assert isinstance(result["status"], str)
+            assert call_args["worker_nodes"] == complex_params["worker_nodes"]
 
     @pytest.mark.asyncio
     async def test_concurrent_tool_calls(self):
@@ -260,138 +208,8 @@ class TestMCPIntegration:
             assert mock_init.call_count == 3
 
     @pytest.mark.asyncio
-    async def test_tool_call_with_complex_parameters(self):
-        """Test tool calls with complex parameter structures."""
-        registry = ToolRegistry(RayManager())
-
-        with patch.object(registry.ray_manager, "init_cluster") as mock_init:
-            mock_init.return_value = {"status": "connected"}
-
-            complex_params = {
-                "num_cpus": 4,
-                "num_gpus": 1,
-                "object_store_memory": 1000000000,
-                "worker_nodes": [
-                    {"num_cpus": 2, "num_gpus": 0, "node_name": "worker-1"},
-                    {"num_cpus": 2, "num_gpus": 1, "node_name": "worker-2"},
-                ],
-                "dashboard_port": 8265,
-                "head_node_host": "127.0.0.1",
-            }
-
-            result = await registry.execute_tool("init_ray", complex_params)
-            assert result["status"] == "connected"
-
-            # Verify complex parameters were passed correctly
-            mock_init.assert_called_once()
-            call_args = mock_init.call_args[1]
-            assert call_args["num_cpus"] == 4
-            assert call_args["worker_nodes"] == complex_params["worker_nodes"]
-
-    @pytest.mark.asyncio
-    async def test_retrieve_logs_tool(self):
-        """Test retrieve_logs tool integration."""
-        registry = ToolRegistry(RayManager())
-
-        with patch.object(registry.ray_manager, "retrieve_logs") as mock_retrieve:
-            mock_retrieve.return_value = {
-                "status": "success",
-                "logs": "Job log line 1\nJob log line 2",
-                "log_type": "job",
-                "identifier": "job_123",
-            }
-
-            result = await registry.execute_tool(
-                "retrieve_logs",
-                {
-                    "identifier": "job_123",
-                    "log_type": "job",
-                    "num_lines": 100,
-                    "include_errors": False,
-                },
-            )
-
-            assert result["status"] == "success"
-            assert "logs" in result
-            assert result["log_type"] == "job"
-            assert result["identifier"] == "job_123"
-
-    @pytest.mark.asyncio
-    async def test_unknown_tool_error(self):
-        """Test error handling for unknown tools."""
-        registry = ToolRegistry(RayManager())
-        result = await registry.execute_tool("unknown_tool", {})
-        assert result["status"] == "error"
-        assert "Unknown tool" in result["message"]
-
-    @pytest.mark.asyncio
-    async def test_parameter_isolation_init_ray(self):
-        """Test that init_ray function only passes expected parameters."""
-        registry = ToolRegistry(RayManager())
-        called_params = {}
-
-        async def fake_init_cluster(**kwargs):
-            called_params.update(kwargs)
-            return {
-                "status": "connected",
-                "message": "Successfully connected to Ray cluster at 127.0.0.1:10001",
-                "cluster_address": "127.0.0.1:10001",
-                "dashboard_url": "http://127.0.0.1:8265",
-                "node_id": "node_123",
-                "job_client_status": "ready",
-            }
-
-        mock = AsyncMock(side_effect=fake_init_cluster)
-        mock.__signature__ = inspect.signature(RayManager.init_cluster)
-
-        with patch.object(RayManager, "init_cluster", mock):
-            result = await registry.execute_tool(
-                "init_ray", {"num_cpus": 4, "num_gpus": 1, "head_node_port": 10001}
-            )
-            expected_params = {"num_cpus", "num_gpus", "head_node_port"}
-            actual_params = set(called_params.keys())
-            unexpected_params = actual_params - expected_params
-            assert (
-                not unexpected_params
-            ), f"Unexpected parameters passed to RayManager: {unexpected_params}"
-            assert called_params["num_cpus"] == 4
-            assert called_params["num_gpus"] == 1
-            assert called_params["head_node_port"] == 10001
-
-    @pytest.mark.asyncio
-    async def test_parameter_isolation_submit_job(self):
-        """Test that submit_job function works correctly with expected parameters."""
-        registry = ToolRegistry(RayManager())
-
-        # Mock the RayManager.submit_job method to return a successful response
-        with patch.object(registry.ray_manager, "submit_job") as mock_submit:
-            mock_submit.return_value = {
-                "status": "success",
-                "result_type": "submitted",
-                "job_id": "job_123",
-            }
-            with patch("ray_mcp.ray_manager.RAY_AVAILABLE", True):
-                with patch("ray_mcp.ray_manager.ray") as mock_ray:
-                    mock_ray.is_initialized.return_value = True
-                    registry.ray_manager._update_state(initialized=True)
-                    result = await registry.execute_tool(
-                        "submit_job",
-                        {
-                            "entrypoint": "python test.py",
-                            "runtime_env": {"pip": ["requests"]},
-                            "job_id": "custom_job_id",
-                        },
-                    )
-                    # Check that the tool executed successfully
-                    assert result["status"] == "success"
-                    assert result["result_type"] == "submitted"
-                    assert result["job_id"] == "job_123"
-                    # Verify the mock was called
-                    mock_submit.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_inspect_job_tool(self):
-        """Test inspect_job tool with different modes."""
+    async def test_inspect_job_modes(self):
+        """Test inspect_job tool with different inspection modes."""
         registry = ToolRegistry(RayManager())
 
         # Mock RayManager method
@@ -438,14 +256,44 @@ class TestMCPIntegration:
             inspect_job_mock.assert_called_with("job_123", "debug")
 
     @pytest.mark.asyncio
-    async def test_init_ray_with_worker_nodes(self):
-        """Test init_ray with worker node configuration."""
+    async def test_retrieve_logs_integration(self):
+        """Test retrieve_logs tool integration."""
         registry = ToolRegistry(RayManager())
-        worker_config = [
-            {"num_cpus": 2, "num_gpus": 1, "node_name": "worker-1"},
-            {"num_cpus": 2, "num_gpus": 0, "node_name": "worker-2"},
-        ]
 
+        with patch.object(registry.ray_manager, "retrieve_logs") as mock_retrieve:
+            mock_retrieve.return_value = {
+                "status": "success",
+                "logs": "Job log line 1\nJob log line 2",
+                "log_type": "job",
+                "identifier": "job_123",
+            }
+
+            result = await registry.execute_tool(
+                "retrieve_logs",
+                {
+                    "identifier": "job_123",
+                    "log_type": "job",
+                    "num_lines": 100,
+                    "include_errors": False,
+                },
+            )
+
+            assert result["status"] == "success"
+            assert "logs" in result
+            assert result["log_type"] == "job"
+            assert result["identifier"] == "job_123"
+
+    @pytest.mark.asyncio
+    async def test_unknown_tool_and_response_format(self):
+        """Test unknown tool handling and response format consistency."""
+        registry = ToolRegistry(RayManager())
+        
+        # Test unknown tool
+        result = await registry.execute_tool("unknown_tool", {})
+        assert result["status"] == "error"
+        assert "Unknown tool" in result["message"]
+
+        # Test response format consistency
         with patch.object(registry.ray_manager, "init_cluster") as mock_init:
             mock_init.return_value = {
                 "status": "connected",
@@ -456,17 +304,12 @@ class TestMCPIntegration:
                 "job_client_status": "ready",
             }
 
-            result = await registry.execute_tool(
-                "init_ray", {"num_cpus": 4, "worker_nodes": worker_config}
-            )
-            assert result["status"] == "connected"
-            mock_init.assert_called_once()
+            result = await registry.execute_tool("init_ray", {"num_cpus": 4})
 
-
-@pytest.mark.integration
-class TestIntegration:
-    # Add more integration tests as needed, using registry.execute_tool or the actual server interface
-    pass
+            # Check response format
+            assert isinstance(result, dict)
+            assert "status" in result
+            assert isinstance(result["status"], str)
 
 
 if __name__ == "__main__":
