+++ tests/test_ray_manager.py
@@ -4,13 +4,23 @@
 import asyncio
 import inspect
 import subprocess
+import threading
 import time
 from typing import Any, Dict, List
 from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
 import pytest
 
-from ray_mcp.ray_manager import RayManager
+from ray_mcp.logging_utils import LogProcessor
+from ray_mcp.ray_manager import (
+    JobClientError,
+    JobConnectionError,
+    JobRuntimeError,
+    JobSubmissionError,
+    JobValidationError,
+    RayManager,
+    RayStateManager,
+)
 from tests.conftest import mock_cluster_startup
 
 
@@ -77,14 +87,6 @@ class TestRayManager:
                 assert result["status"] == "not_running"
 
     @pytest.mark.asyncio
-    async def test_stop_cluster_ray_unavailable(self, manager):
-        """Test stopping when Ray is not available."""
-        with patch("ray_mcp.ray_manager.RAY_AVAILABLE", False):
-            result = await manager.stop_cluster()
-            assert result["status"] == "error"
-            assert "Ray is not available" in result["message"]
-
-    @pytest.mark.asyncio
     async def test_cleanup_head_node_process(self, manager):
         """Test enhanced cleanup of head node process with a real subprocess."""
         import subprocess
@@ -122,15 +124,6 @@ class TestRayManager:
                 )
                 assert result is None
 
-    @pytest.mark.asyncio
-    async def test_initialize_job_client_not_available(self, manager):
-        """Test job client initialization when JobSubmissionClient is not available."""
-        with patch("ray_mcp.ray_manager.JobSubmissionClient", None):
-            result = await manager._initialize_job_client_with_retry(
-                "http://127.0.0.1:8265"
-            )
-            assert result is None
-
     def test_address_format_validation(self):
         """Test that address format validation works correctly."""
         manager = RayManager()
@@ -148,6 +141,234 @@ class TestRayManager:
             assert isinstance(address, str)
             assert ":" in address
             assert address.split(":")[1].isdigit()
+
+
+@pytest.mark.fast
+class TestExceptionHandling:
+    """Test exception handling for job submission methods."""
+
+    @pytest.fixture
+    def ray_manager(self):
+        """Create a RayManager instance for testing."""
+        manager = RayManager()
+        # Set internal state to avoid actual Ray setup
+        manager._update_state(initialized=True)
+        manager._job_client = Mock()
+        return manager
+
+    def _patch_is_initialized(self, manager):
+        """Helper method to patch is_initialized property."""
+        original_property = type(manager).is_initialized
+        type(manager).is_initialized = property(lambda self: True)
+        return original_property
+
+    def _restore_is_initialized(self, manager, original_property):
+        """Helper method to restore the original is_initialized property."""
+        type(manager).is_initialized = original_property
+
+    @pytest.mark.asyncio
+    async def test_submit_job_handles_validation_error(self, ray_manager):
+        """Test that submit_job properly handles validation errors."""
+        original_property = self._patch_is_initialized(ray_manager)
+        try:
+            # Test with empty entrypoint
+            result = await ray_manager.submit_job("")
+            assert result["status"] == "error"
+            assert "Entrypoint cannot be empty" in result["message"]
+
+            # Test with whitespace-only entrypoint
+            result = await ray_manager.submit_job("   ")
+            assert result["status"] == "error"
+            assert "Entrypoint cannot be empty" in result["message"]
+        finally:
+            self._restore_is_initialized(ray_manager, original_property)
+
+    @pytest.mark.asyncio
+    async def test_cancel_job_handles_validation_error(self, ray_manager):
+        """Test that cancel_job properly handles validation errors."""
+        original_property = self._patch_is_initialized(ray_manager)
+        try:
+            # Test with empty job_id
+            result = await ray_manager.cancel_job("")
+            assert result["status"] == "error"
+            assert "Job ID cannot be empty" in result["message"]
+
+            # Test with whitespace-only job_id
+            result = await ray_manager.cancel_job("   ")
+            assert result["status"] == "error" 
+            assert "Job ID cannot be empty" in result["message"]
+        finally:
+            self._restore_is_initialized(ray_manager, original_property)
+
+    def test_custom_exception_hierarchy(self):
+        """Test that custom exception hierarchy is properly defined."""
+        # Test that JobSubmissionError is the base class
+        assert issubclass(JobConnectionError, JobSubmissionError)
+        assert issubclass(JobValidationError, JobSubmissionError)
+        assert issubclass(JobRuntimeError, JobSubmissionError)
+        assert issubclass(JobClientError, JobSubmissionError)
+
+        # Test that exceptions can be instantiated
+        connection_error = JobConnectionError("Connection failed")
+        validation_error = JobValidationError("Invalid parameters")
+        runtime_error = JobRuntimeError("Runtime error")
+        client_error = JobClientError("Client not initialized")
+
+        assert str(connection_error) == "Connection failed"
+        assert str(validation_error) == "Invalid parameters"
+        assert str(runtime_error) == "Runtime error"
+        assert str(client_error) == "Client not initialized"
+
+
+@pytest.mark.fast
+class TestStreamingLogs:
+    """Test cases for streaming log retrieval functionality."""
+
+    @pytest.fixture
+    def ray_manager(self):
+        """Create a RayManager instance for testing."""
+        manager = RayManager()
+        manager._update_state(initialized=True)
+        manager._job_client = Mock()
+        manager._ensure_initialized = lambda: None
+        return manager
+
+    def test_validate_log_parameters_valid(self):
+        result = LogProcessor.validate_log_parameters(100, 10)
+        assert result is None
+
+    def test_validate_log_parameters_invalid_num_lines(self):
+        result = LogProcessor.validate_log_parameters(0, 10)
+        assert result["status"] == "error"
+        assert "num_lines must be positive" in result["message"]
+
+        result = LogProcessor.validate_log_parameters(15000, 10)
+        assert result["status"] == "error"
+        assert "num_lines cannot exceed 10000" in result["message"]
+
+    def test_validate_log_parameters_invalid_max_size(self):
+        result = LogProcessor.validate_log_parameters(100, 0)
+        assert result["status"] == "error"
+        assert "max_size_mb must be between 1 and 100" in result["message"]
+
+        result = LogProcessor.validate_log_parameters(100, 150)
+        assert result["status"] == "error"
+        assert "max_size_mb must be between 1 and 100" in result["message"]
+
+    def test_truncate_logs_to_size_within_limit(self):
+        logs = "This is a test log\nwith multiple lines\nbut small size"
+        result = LogProcessor.truncate_logs_to_size(logs, 10)
+        assert result == logs
+
+    def test_truncate_logs_to_size_exceeds_limit(self):
+        large_log = "x" * (2 * 1024 * 1024)
+        result = LogProcessor.truncate_logs_to_size(large_log, 1)
+        max_allowed = 1024 * 1024 + 1024
+        assert len(result.encode("utf-8")) <= max_allowed
+        assert "... (truncated at 1MB limit)" in result
+
+    def test_stream_logs_with_limits_string_input(self):
+        logs = "line1\nline2\nline3\nline4\nline5"
+        result = LogProcessor.stream_logs_with_limits(logs, max_lines=3, max_size_mb=1)
+        lines = result.split("\n")
+        assert len(lines) == 4
+        assert lines[0] == "line1"
+        assert lines[1] == "line2"
+        assert lines[2] == "line3"
+        assert "... (truncated at 3 lines)" in lines[3]
+
+    def test_analyze_job_logs_with_streaming(self, ray_manager):
+        """Test log analysis with streaming approach."""
+        logs = "This is a normal log\nThis is an error log\nThis is an exception\nNormal log again"
+        result = ray_manager._analyze_job_logs(logs)
+
+        assert result["error_count"] == 2
+        assert len(result["errors"]) == 2
+        assert "error" in result["errors"][0].lower()
+
+
+@pytest.mark.fast
+class TestStateManagement:
+    """Test core state management functionality."""
+
+    def test_ray_state_manager_initialization(self):
+        """Test RayStateManager initialization."""
+        state_manager = RayStateManager()
+        state = state_manager.get_state()
+
+        assert state["initialized"] is False
+        assert state["cluster_address"] is None
+        assert state["dashboard_url"] is None
+        assert state["job_client"] is None
+
+    def test_state_update_and_reset(self):
+        """Test state updates and reset."""
+        state_manager = RayStateManager()
+
+        # Update state
+        state_manager.update_state(
+            initialized=True,
+            cluster_address="ray://localhost:10001",
+            dashboard_url="http://localhost:8265",
+        )
+
+        state = state_manager.get_state()
+        assert state["initialized"] is True
+        assert state["cluster_address"] == "ray://localhost:10001"
+        assert state["dashboard_url"] == "http://localhost:8265"
+
+        # Reset state
+        state_manager.reset_state()
+        state = state_manager.get_state()
+        assert state["initialized"] is False
+        assert state["cluster_address"] is None
+
+    def test_ray_manager_state_integration(self):
+        """Test RayManager state management integration."""
+        manager = RayManager()
+
+        # Test initial state
+        assert manager.is_initialized is False
+        assert manager.cluster_address is None
+        assert manager.dashboard_url is None
+
+        # Test state update
+        manager._update_state(
+            initialized=True,
+            cluster_address="ray://localhost:10001", 
+            dashboard_url="http://localhost:8265"
+        )
+        assert manager.is_initialized is True
+        assert manager.cluster_address == "ray://localhost:10001"
+        assert manager.dashboard_url == "http://localhost:8265"
+
+    def test_thread_safety_basic(self):
+        """Test basic thread safety of state management."""
+        state_manager = RayStateManager()
+        results = []
+
+        def update_state_thread(thread_id):
+            for i in range(50):
+                state_manager.update_state(
+                    initialized=True,
+                    cluster_address=f"ray://localhost:{10001 + thread_id}",
+                )
+                state = state_manager.get_state()
+                results.append((thread_id, state["initialized"]))
+                time.sleep(0.001)
+
+        # Start multiple threads
+        threads = []
+        for i in range(3):
+            thread = threading.Thread(target=update_state_thread, args=(i,))
+            threads.append(thread)
+            thread.start()
+
+        for thread in threads:
+            thread.join()
+
+        # Verify no exceptions occurred
+        assert len(results) == 150  # 3 threads * 50 updates each
 
 
 if __name__ == "__main__":
